+++ tests/base/src/test/resources/org/keycloak/tests/admin/identityprovider/saml-idp-metadata-different-bindings.xml
