+++ tests/base/src/test/resources/org/keycloak/tests/admin/identityprovider/saml-idp-metadata_utf8_bom.xml
